Add `:args` to Docker `build` command.
When the docker build command is issued this plugin will take in an
`:args` argument that specifies the build arguments to be passed to
Docker.

Each argument is specified as a map containing `:name` and `:value`
keys. For example `[{:name "foo" :value "bar"} {:name "a" :value "b"}]
will be turned into `--build-arg foo=bar --build-arg a=b` on the
docker build command executed by lein.